Make maxNumberOfNodes not static. Version 0.2 now.
In order to circumvent issues with the java compiler, which replaces constant values in the byte code and removes it from ifs. The maximum number of nodes is stored inside a Long variable. The Long type is used instead of long (primitive) because removing a private static variable leaves it's it's initialization inside the static block. Therefore re-assigning a value to that variable had no affect. It seems to be the trade-offs when fiddling with byte code after optimization. Using a Long instance circumvented those static initialization issues.

+Increasing up to version 0.2.